moved some class non-essentials to utils mod
