+++ tests/xontrib_chatgpt/test_chatmanager.py
@@ -32,28 +32,6 @@ def cm():
     return ChatManager()
 
 
-@pytest.fixture
-def sys_msgs():
-    l = """[
-    {'role': 'system', 'content': 'Hello'},
-    {'role': 'system', 'content': 'Hi there!'},
-    ]
-    """
-
-    d = '{"content": "Hello"}'
-
-    y = dedent(
-        """
-    - role: system
-      content: Hello
-    - role: system
-      content: Hi there!
-    """
-    )
-
-    return l, d, y
-
-
 def test_update_inst_dict(xession, cm):
     insts = [
         ("test", ChatGPT("test")),
@@ -246,22 +224,6 @@ def test_cli(xession, cm, action, args, expected, monkeypatch):
     )
     cm(args)
     assert getattr(cm, f"_{action}") == expected
-
-
-def test_convert_to_sys(xession, sys_msgs):
-    l, d, y = sys_msgs
-    res = convert_to_sys(l)
-    assert res == [
-        {"role": "system", "content": "Hello"},
-        {"role": "system", "content": "Hi there!"},
-    ]
-    res = convert_to_sys(d)
-    assert res == [{"role": "system", "content": "Hello"}]
-    res = convert_to_sys(y)
-    assert res == [
-        {"role": "system", "content": "Hello"},
-        {"role": "system", "content": "Hi there!"},
-    ]
 
 
 def test_edit(xession, cm, cm_events):
